# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Plan Management API
  version: 1.0.0
  description: 주간 계획표 관리 서비스의 API 문서입니다.
servers:
  - url: http://localhost:4000/api
    description: Local server
paths:
  /auth/signup:
    post:
      summary: "회원가입"
      description: "새로운 사용자를 등록합니다. (관리자 권한 필요)"
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "test_user"
                password:
                  type: string
                  example: "123456"
                user_name:
                  type: string
                  example: "홍길동"
                number:
                  type: string
                  example: "01012345678"
                role:
                  type: string
                  example: "사원"
      responses:
        "201":
          description: "회원가입 성공"
        "400":
          description: "입력값 오류 또는 중복 ID"
        "500":
          description: "서버 오류"
  /auth/login:
    post:
      summary: "로그인"
      description: "아이디와 비밀번호로 로그인 후 토큰 발급."
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "로그인 성공 (accessToken 및 isAdmin 반환)"
        "400":
          description: "아이디 또는 비밀번호 오류"
        "500":
          description: "서버 오류"
...
  /auth/refresh:
    post:
      summary: "Access Token 갱신"
      tags: [Auth]
      responses:
        "200":
          description: "토큰 재발급 성공"
        "403":
          description: "Refresh Token 오류"
  /auth/logout:
    post:
      summary: "로그아웃"
      tags: [Auth]
      responses:
        "200":
          description: "로그아웃 성공"
        "500":
          description: "서버 오류"
  /auth:
    delete:
      summary: "유저 삭제"
      description: "이름으로 유저를 삭제합니다. (관리자 전용)"
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
      responses:
        "200":
          description: "삭제 성공"
        "400":
          description: "입력 누락"
        "404":
          description: "유저 없음"
        "500":
          description: "서버 오류"
  /auth/update:
    patch:
      summary: "유저 정보 수정"
      description: "현재 로그인한 사용자의 정보를 수정합니다."
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                  example: "김유진"
                number:
                  type: string
                  example: "01012345678"
                role:
                  type: string
                  example: "사원"
      responses:
        "200":
          description: "수정 성공"
        "400":
          description: "잘못된 요청 또는 수정할 정보 없음"
        "404":
          description: "사용자 없음"
        "500":
          description: "서버 오류"
  /auth/password:
    patch:
      summary: "비밀번호 변경"
      description: "현재 로그인한 사용자가 비밀번호를 변경합니다."
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: "123456"
                newPassword:
                  type: string
                  example: "newpassword123"
      responses:
        "200":
          description: "비밀번호 변경 성공"
        "400":
          description: "입력 누락 또는 형식 오류"
        "401":
          description: "현재 비밀번호가 일치하지 않음"
        "404":
          description: "사용자 없음"
        "500":
          description: "서버 오류"
...